+++ internal/bloblang/parser/query_function_parser.go
@@ -231,7 +231,7 @@ func variableLiteralParser() Func {
 
 var errNoRoot = errors.New("unable to reference the `root` of your mapped document within a query. This feature will be introduced soon, but in the meantime in order to use a mapped value multiple times use variables (https://www.benthos.dev/docs/guides/bloblang/about#variables). If instead you wish to refer to a field `root` from your input document use `this.root`")
 
-func fieldLiteralRootParser() Func {
+func fieldLiteralRootParser(pCtx Context) Func {
 	fieldPathParser := Expect(
 		JoinStringPayloads(
 			UntilFail(
@@ -260,7 +260,11 @@ func fieldLiteralRootParser() Func {
 		} else if path == "root" {
 			return Fail(NewFatalError(input, errNoRoot), input)
 		} else {
-			fn = query.NewFieldFunction(path)
+			if pCtx.HasNamedContext(path) {
+				fn = query.NewNamedContextFieldFunction(path, "")
+			} else {
+				fn = query.NewFieldFunction(path)
+			}
 		}
 
 		return Success(fn, res.Remaining)