+++ internal/bloblang/query/target.go
@@ -49,20 +49,47 @@ func methodTargetPaths(target, method Function) func(ctx TargetsContext) (Target
 	}
 }
 
-// TargetsContext provides access to a range of query targets for functions to
-// reference when determining their targets.
+// TargetsContext describes the current Bloblang execution environment from the
+// perspective of a particular query function in a way that allows it to
+// determine which values it is targetting and the origins of those values.
+//
+// The environment consists of named maps that are globally accessible, the
+// current value that is being executed upon by methods (when applicable), the
+// general (main) context (referenced by the keyword `this`) and any other named
+// contexts accessible at this point.
+//
+// Since it's possible for any query function to reference and return multiple
+// target candidates (match expressions, etc) then each context and the current
+// value are lists of paths, each being a candidate at runtime.
 type TargetsContext struct {
-	Maps          map[string]Function
-	CurrentValues []TargetPath
+	Maps map[string]Function
 
-	mainContext  []TargetPath
-	namedContext *namedContextPath
+	currentValues []TargetPath
+	mainContext   []TargetPath
+	prevContext   *prevContextPath
+	namedContext  *namedContextPath
+}
+
+type prevContextPath struct {
+	paths []TargetPath
+	next  *prevContextPath
 }
 
 type namedContextPath struct {
 	name  string
 	paths []TargetPath
 	next  *namedContextPath
+}
+
+// Value returns the current value of the targets context, which is the path(s)
+// being executed upon by methods.
+func (ctx TargetsContext) Value() []TargetPath {
+	return ctx.currentValues
+}
+
+// MainContext returns the path of the main context.
+func (ctx TargetsContext) MainContext() []TargetPath {
+	return ctx.mainContext
 }
 
 // NamedContext returns the path of a named context if it exists.
@@ -77,24 +104,50 @@ func (ctx TargetsContext) NamedContext(name string) []TargetPath {
 	return nil
 }
 
-// MainContext returns the path of the main context.
-func (ctx TargetsContext) MainContext() []TargetPath {
-	return ctx.mainContext
-}
-
-// WithMainContext returns a TargetsContext with a new main context path.
-func (ctx TargetsContext) WithMainContext(paths []TargetPath) TargetsContext {
-	ctx.mainContext = paths
+// WithValues returns a targets context where the current value being executed
+// upon by methods is set to something new.
+func (ctx TargetsContext) WithValues(paths []TargetPath) TargetsContext {
+	ctx.currentValues = paths
 	return ctx
 }
 
-// WithNamedContext returns a TargetsContext with a named value path.
-func (ctx TargetsContext) WithNamedContext(name string, paths []TargetPath) TargetsContext {
+// WithValuesAsContext returns a targets context where the current value being
+// executed upon by methods is now the main context. This happens when a query
+// function is executed as a method, or within branches of match expressions.
+func (ctx TargetsContext) WithValuesAsContext() TargetsContext {
+	ctx.prevContext = &prevContextPath{
+		paths: ctx.mainContext,
+		next:  ctx.prevContext,
+	}
+	ctx.mainContext = ctx.currentValues
+	ctx.currentValues = nil
+	return ctx
+}
+
+// WithContextAsNamed moves the latest context into a named context and returns
+// the context prior to that one to the main context. This is a way for named
+// context mappings to correct the contexts so that the child query function
+// returns the right paths.
+func (ctx TargetsContext) WithContextAsNamed(name string) TargetsContext {
 	previous := ctx.namedContext
 	ctx.namedContext = &namedContextPath{
 		name:  name,
-		paths: paths,
+		paths: ctx.mainContext,
 		next:  previous,
+	}
+	if ctx.prevContext != nil {
+		ctx.mainContext = ctx.prevContext.paths
+		ctx.prevContext = ctx.prevContext.next
+	}
+	return ctx
+}
+
+// PopContext returns a targets context with the latest context dropped and the
+// previous (when applicable) returned.
+func (ctx TargetsContext) PopContext() TargetsContext {
+	if ctx.prevContext != nil {
+		ctx.mainContext = ctx.prevContext.paths
+		ctx.prevContext = ctx.prevContext.next
 	}
 	return ctx
 }