+++ internal/docs/bloblang.go
@@ -230,16 +230,12 @@ Methods provide most of the power in Bloblang as they allow you to augment value
 
 ` + "```coffee" + `
 root.doc.id = this.thing.id.string().catch(uuid_v4())
-root.doc.reduced_nums = this.thing.nums.map_each(
-  if this < 10 {
-    deleted()
-  } else {
-    this - 10
-  }
-)
-root.has_good_taste = ["pikachu","mewtwo","magmar"].contains(
-  this.user.fav_pokemon
-)
+root.doc.reduced_nums = this.thing.nums.map_each(num -> if num < 10 {
+  deleted()
+} else {
+  num - 10
+})
+root.has_good_taste = ["pikachu","mewtwo","magmar"].contains(this.user.fav_pokemon)
 ` + "```" + `
 
 {{if gt (len .General) 0 -}}